feat: 답변 작성자가 아니라면 수정/삭제 불가능

+++ src/main/java/proj/chat/domain/controller/CommentController.java
@@ -2,6 +2,7 @@ package proj.chat.domain.controller;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
@@ -12,6 +13,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.server.ResponseStatusException;
 import proj.chat.domain.dto.comment.CommentResponseDto;
 import proj.chat.domain.dto.comment.CommentSaveRequestDto;
 import proj.chat.domain.dto.comment.CommentUpdateRequestDto;
@@ -32,17 +34,17 @@ public class CommentController {
     public String save(@PathVariable("postId") Long postId,
             @Validated @ModelAttribute CommentSaveRequestDto requestDto,
             BindingResult bindingResult, Authentication authentication, Model model) {
-    
+        
         if (bindingResult.hasErrors()) {
             model.addAttribute("post", postService.findById(postId));
             return "post/detail";
         }
-    
+        
         CustomUserDetails principal = (CustomUserDetails) authentication.getPrincipal();
-    
+        
         requestDto.setPostId(postId);
         Long savedId = commentService.save(requestDto, principal.getMember().getEmail());
-    
+        
         return "redirect:/post/" + postId;
     }
     
@@ -50,8 +52,16 @@ public class CommentController {
     @GetMapping("/comment/{id}/update")
     public String updateForm(@PathVariable("id") Long id,
             @ModelAttribute("commentUpdateRequestDto") CommentUpdateRequestDto requestDto,
-            Model model) {
-    
+            Authentication authentication, Model model) {
+        
+        CustomUserDetails principal = (CustomUserDetails) authentication.getPrincipal();
+        
+        CommentResponseDto findCommentDto = commentService.findById(id);
+        if (!findCommentDto.getMemberEmail().equals(principal.getMember().getEmail())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "수정 권한이 없습니다");
+        }
+        
+        requestDto.setContent(findCommentDto.getContent());
         model.addAttribute("commentId", id);
         return "comment/update";
     }
@@ -60,28 +70,40 @@ public class CommentController {
     @PostMapping("/comment/{id}/update")
     public String update(@PathVariable("id") Long id,
             @Validated @ModelAttribute CommentUpdateRequestDto requestDto,
-            BindingResult bindingResult) {
-    
+            BindingResult bindingResult, Authentication authentication) {
+        
         if (bindingResult.hasErrors()) {
-            
-            bindingResult.rejectValue("content", "답변 수정에 실패했습니다");
             return "comment/update";
         }
-    
+        
+        CustomUserDetails principal = (CustomUserDetails) authentication.getPrincipal();
+        
+        CommentResponseDto findCommentDto = commentService.findById(id);
+        if (!findCommentDto.getMemberEmail().equals(principal.getMember().getEmail())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "수정 권한이 없습니다");
+        }
+        
         Long updatedId = commentService.update(id, requestDto);
         CommentResponseDto findComment = commentService.findById(updatedId);
-    
+        
         return "redirect:/post/" + findComment.getPostId();
     }
     
     @PreAuthorize("isAuthenticated()")
-    @PostMapping("/comment/{id}/delete")
-    public String delete(@PathVariable("id") Long id) {
-    
+    @GetMapping("/comment/{id}/delete")
+    public String delete(@PathVariable("id") Long id, Authentication authentication) {
+        
+        CustomUserDetails principal = (CustomUserDetails) authentication.getPrincipal();
+        
+        CommentResponseDto findCommentDto = commentService.findById(id);
+        if (!findCommentDto.getMemberEmail().equals(principal.getMember().getEmail())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "삭제 권한이 없습니다");
+        }
+        
         Long postId = commentService.findById(id).getPostId();
-    
+        
         Long deletedId = commentService.delete(id);
-    
+        
         return "redirect:/post/" + postId;
     }
 }
